add github action yml file, consolidate files on server side, add authentication on server side
