+++ server/src/models/index.ts
@@ -1,6 +1,6 @@
 // Exports for all models from here
 import Cryptid from './Cryptid.js'
 import ForumPost from './ForumPost.js'
-import Sightings from './Sightings.js'
+//import Sightings from './Sightings.js'
 
-export default { Cryptid, ForumPost, Sightings } 
+export default { Cryptid, ForumPost }
