+++ server/src/schemas/index.ts
@@ -1,0 +1,4 @@
+import typeDefs from "./typeDefs.js";
+import resolvers from './resolvers.js'
+
+export { typeDefs, resolvers }
