--- server/src/schemas/resolvers/cryptidResolvers.ts
@@ -1,1 +1,0 @@
-// Handles GraphQL queries/mutations related to cryptids and sightings
